fix: swap jotai
