+++ src/hooks/useApproveCallback/hedera.ts
@@ -6,7 +6,7 @@ import { ZERO_ADDRESS } from 'src/constants';
 import { ROUTER_ADDRESS, ROUTER_DAAS_ADDRESS } from 'src/constants/address';
 import { useTokenAllowance } from 'src/data/Allowances';
 import { useHederaTotalSupply } from 'src/data/TotalSupply';
-import { Field } from 'src/state/pswap/actions';
+import { Field } from 'src/state/pswap/atom';
 import { useHasPendingApproval, useTransactionAdder } from 'src/state/ptransactions/hooks';
 import { useIsApprovingInfinite } from 'src/state/puser/hooks';
 import { fetchHederaPGLToken } from 'src/state/pwallet/hooks/hedera';
