+++ src/state/index.ts
@@ -6,13 +6,11 @@ import { load, save } from 'redux-localstorage-simple';
 import pbridge from './pbridge/reducer';
 import plists from './plists/reducer';
 import pmulticall from './pmulticall/reducer';
-import pswap from './pswap/reducer';
 import puser from './puser/reducer';
 
 export const PANGOLIN_PERSISTED_KEYS: string[] = ['puser', 'plists'];
 
 export const pangolinReducers = {
-  pswap,
   plists,
   pmulticall,
   puser,
